adversarial_attack_imagenet/requirements.txt
cycler==0.10.0
eagerpy==0.29.0
future==0.18.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.7.4.3
